backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
mysqlclient
gunicorn
Pillow
asgiref
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
framer-motion
react-intersection-observer
request